+++ Base/src/util/timer.hpp
@@ -1,10 +1,12 @@
 #ifndef TIMER_H
 #define TIMER_H
 
+#include <cstdint>
+
 class Timer
 {
 public:
-	static long long time_now();
+	static int64_t time_now();
 	static void sleep(int milliseconds);
 };
